Leafdown R6 Class

This class acts as a wrapper around the leaflet map. It allows the user to: - select shapes - drill down on these selected shapes - drill up - convenient functions to connect the map to graphs

Active bindings

spdfs_list

The spdfs of all map levels. This is set in initialize and cannot be changed afterwards.

map_output_id

The id from the shiny-ui used in the leafletOutput("<<id>>"). Used to observe for _shape_click events.

curr_data

The spdf-metadata from the currently used shapes AND the corresponding values.

curr_map_level

The current map level. This corresponds to the position of the shapes in the spdfs_list. (i.e The highest-level is 1, the next is 2 and so on...)

curr_selection

The selected shapes of the current level. They will be highlighted on the map. Calling drill_down will drill down on these selected shapes.

curr_spdf

The spdfs of the current map level. curr_poly_ids The ids of all polygons of the current map level.

Methods

Public methods

Leafdown$new()

Leafdown$draw_leafdown()

Leafdown$get_current_metadata()

Leafdown$add_data()

Leafdown$drill_down()

Leafdown$drill_up()

Leafdown$clone()

Method new()

Initializes the observer for the maps _shape_click events. This is needed for the selection. Once a shape is clicked, it is added to the .curr_selection (or removed from it). The outline of selected shapes is highlighted via the showGroup (hideGroup) functions.

Usage

Leafdown$new(spdfs_list, map_output_id, input)

Arguments

spdfs_list

The spdfs of all map levels. This cannot be changed later

map_output_id

The id from the shiny-ui used in the leafletOutput("<<id>>"). Used to observe for _shape_click events.

input

The input from the shiny app

Method draw_leafdown()

Draws the leaflet map on the current map level. All unselected parents will be drawn in gray.

Usage

Leafdown$draw_leafdown(...)

Arguments

...

Additional arguments given to leaflet::addPolygons

Method get_current_metadata()

Returns the metadata of the shapes from the current maplevel. This may differ from what is displayed if drill_down was called but draw_leafdown wasn't (yet)

Usage

Leafdown$get_current_metadata()

Returns

The current meta-data

Method add_data()

Sets the data of the current shapes. This includes the meta-data AND the values given by the user. These values can be used to draw differently colored shapes.

Usage

Leafdown$add_data(data)

Arguments

data

The new data existing of the meta-data and the values to display in the map(color)

Method drill_down()

Drills down to the lower level if: - there is a lower level (for now there are only two levels) - at least one shape is selected to drill down on !This will not redraw the map! Also call add_data to add data for the new level and then draw_leafdown to redraw the map on the new level

Usage

Leafdown$drill_down()

Method drill_up()

Drills up to the higher level if: - there is a higher level (for now there are only two levels) !This will not redraw the map! Also call add_data to add data for the new level and then draw_leafdown to redraw the map on the new level

Usage

Leafdown$drill_up()

Method clone()

The objects of this class are cloneable with this method.

Usage

Leafdown$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.