Leafdown R6 Class

This class acts as a wrapper around the leaflet map. A leaflet map allows the user of the shiny app to:

select shapes / regions

drill down on these selected shapes

drill up

The active bindings of the class provide convenient interfaces between the leafdown map and other elements in the shiny app e.g. graphs, tables etc. A leafdown object can only be used in a shiny app.

Active bindings

spdfs_list

The spdfs of all map levels. This is set in initialize and cannot be changed afterwards. At the moment only two map levels are possible.

map_output_id

The id from the shiny ui used in the leafletOutput("<<id>>"). Used to observe for _shape_click events.

curr_data

The metadata and (if available) the corresponding values of all currently displayed shapes.

curr_sel_data

A reactiveValue containing a data.frame with the metadata and (if available) the corresponding values of all currently selected shapes.

curr_map_level

Index of the current map level. This corresponds to the position of the shapes in the spdfs_list. (i.e The highest-level is 1, the next is 2 and so on...). At the moment only two map levels are possible. curr_sel_ids The ids of the selected shapes of the current level. They will be highlighted on the map. Calling drill_down, the drill down functionality is executed for these shapes.

curr_spdf

The spdfs of the current map level. curr_poly_ids The ids of all polygons of the current map level. selected_parents The selected spdf shapes from the higher level. (Subset of spdfs_list) unselected_parents All spdf shapes from the higher level which are not selected. They will be drawn in gray. (Subset of spdfs_list)

Methods

Public methods

Leafdown$new()

Leafdown$draw_leafdown()

Leafdown$get_current_metadata()

Leafdown$add_data()

Leafdown$drill_down()

Leafdown$drill_up()

Leafdown$toggle_shape_select()

Leafdown$clone()

Method new()

Initializes the observer for the maps _shape_click events. This is needed for the shape selection. Once a shape is clicked, it is added to (or removed from) .curr_sel_ids. The outline of selected shapes is highlighted via the showGroup (hideGroup) functions.

Initializes the leafdown object. This will not draw the map. First add data and then call draw_leafdown to draw the map.

Usage

Leafdown$new(spdfs_list, map_output_id, input)

Arguments

spdfs_list

The spdfs of all map levels. This cannot be changed later

map_output_id

The id from the shiny-ui used in the leafletOutput("<<id>>"). Used to observe for _shape_click events.

input

The input from the shiny app

Method draw_leafdown()

Draws the leaflet map on the current map level. All unselected parents will be drawn in gray.

Usage

Leafdown$draw_leafdown(...)

Arguments

...

Additional arguments given to leaflet::addPolygons

Method get_current_metadata()

Returns the metadata of the shapes from the current maplevel. This may differ from what is displayed if drill_down has been called but draw_leafdown has not been (yet).

Usage

Leafdown$get_current_metadata()

Returns

The current meta-data

Method add_data()

Adds the data to the currently displayed shapes. This includes the meta-data AND the values to be visualized in the map.

Usage

Leafdown$add_data(data)

Arguments

data

The new data existing of the meta-data and the values to display in the map(color)

Method drill_down()

Drills down to the lower level if: - there is a lower level (for now there are only two levels) - at least one shape is selected to drill down on !This will not redraw the map! Also call add_data to add data for the new level and then draw_leafdown to redraw the map on the new level

Usage

Leafdown$drill_down()

Method drill_up()

Drills up to the higher level if: - there is a higher level (for now there are only two levels) !This will not redraw the map! Also call add_data to add data for the new level and then draw_leafdown to redraw the map on the new level

Usage

Leafdown$drill_up()

Method toggle_shape_select()

Selects the shape with the given shape id, or unselects it if it was already selected.

Usage

Leafdown$toggle_shape_select(shape_id)

Arguments

shape_id

the id of the shape to select, has to be a character and in the current map-level.

Method clone()

The objects of this class are cloneable with this method.

Usage

Leafdown$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.